requirements.txt
python-dotenv==1.0.1
gspread==6.1.4
google-api-python-client==2.149.0
google-auth==2.36.0
anthropic==0.39.0
requests==2.32.3
jinja2==3.1.4
flask==3.1.0
